refactor(geometry): group all geometry utilities into a new package
