+++ .scalafmt.conf
@@ -37,7 +37,7 @@ rewrite.sortModifiers.order = [
 ]
 rewrite.scala3.convertToNewSyntax = true
 rewrite.scala3.insertEndMarkerMinLines = 20
-rewrite.scala3.removeEndMarkerMaxLines = 10
+rewrite.scala3.removeEndMarkerMaxLines = 19
 rewrite.scala3.removeOptionalBraces = oldSyntaxToo
 
 # Space
