+++ src/main/scala/io/github/srs/utils/collision/Types.scala
@@ -2,85 +2,7 @@ package io.github.srs.utils.collision
 
 import io.github.srs.model.entity.{ Orientation, Point2D, ShapeType }
 import io.github.srs.utils.*
-
-/**
- * Represents a 2D vector in the collision system.
- */
-type Vector2D = Point2D
-
-/**
- * Companion object for Vector2D, providing utility methods for vector operations.
- */
-object Vector2D:
-
-  /**
-   * Creates a new Vector2D instance from the given x and y coordinates.
-   * @param x
-   *   the x-coordinate of the vector
-   * @param y
-   *   the y-coordinate of the vector
-   * @return
-   *   a new Vector2D instance representing the vector in 2D space
-   */
-  def apply(x: Double, y: Double): Vector2D = Point2D(x, y)
-
-  extension (v: Vector2D)
-
-    /**
-     * Rotates the vector by the specified angle.
-     * @param angle
-     *   the angle to rotate the vector by
-     * @return
-     *   a new Vector2D instance representing the rotated vector
-     */
-    def rotate(angle: Orientation): Vector2D =
-      import Point2D.*
-      val radians = angle.toRadians
-      val cos = Math.cos(radians)
-      val sin = Math.sin(radians)
-      Point2D(
-        v.x * cos - v.y * sin,
-        v.x * sin + v.y * cos,
-      )
-
-    /**
-     * Projects the vector onto a line defined by its origin and direction.
-     * @param line
-     *   the line onto which the vector will be projected
-     * @return
-     *   a new Vector2D instance representing the projection of the vector onto the line
-     */
-    def project(line: Line): Vector2D =
-      import Point2D.*
-      val lineDirection = line.direction
-      val dotProduct = line.direction.x * (v.x - line.origin.x) + line.direction.y * (v.y - line.origin.y)
-      Vector2D(
-        line.origin.x + (dotProduct * lineDirection.x),
-        line.origin.y + (dotProduct * lineDirection.y),
-      )
-  end extension
-end Vector2D
-
-/**
- * Represents a line in the collision system, defined by an origin point and a direction vector.
- */
-private[collision] type Line = (origin: Point2D, direction: Point2D)
-
-/**
- * Companion object for Line, providing utility methods for creating and manipulating lines.
- */
-private[collision] object Line:
-
-  /**
-   * Creates a new Line instance from the given origin and direction points.
-   * @param origin
-   *   the starting point of the line
-   * @param direction
-   *   the direction vector of the line
-   * @return
-   *   a new Line instance representing the line in 2D space
-   */
-  def apply(origin: Point2D, direction: Point2D): Line = (origin, direction)
+import io.github.srs.utils.geometry2d.{ Line, Vector2D }
 
 /**
  * Represents an axis in the collision system, defined by two lines: one for the x-axis and one for the y-axis.
